Move validation code over to separate module
